Use godep to add go dependencies.
This is a work in progress to lock down the gomon
dependencies.

To install `godef`:

  go get -u code.google.com/p/rog-go/exp/cmd/godef

However, to get the `godef` command to work, I had
to install these dependencies:

  go get github.com/onsi/gomega
  go get github.com/onsi/ginkgo

This diff represents all the latest versions at the time
of the commit.  This diff was created by using
the following command:

  godep save

Note:  This does fix the build.  I will update the
dependency versions in the Godeps.json file to the
appropriate version.