Add godeps to revision lock cloudwatch library.
This change introduces `godep` to mark the
dependencies on external libraries.

To install `godep`:

  go get github.com/kr/godep

To generate the initial Godeps file run:

  godep save -copy=false

This creates a dependency JSON file.  I edited
the verison SHA's for two libraries:

  https://github.com/crowdmob/goamz/aws
  https://github.com/crowdmob/goamz/cloudwatch

I set the SHA to this version:

  bfa553e3b7d6416a36cb81ff5f379acfb9b93f0d

Using this SHA for the libaries, I was able
to get the build to work with the following command:

  godep go build